% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlerate.R
\name{xlerate}
\alias{xlerate}
\title{Create a function from an excel workbook}
\usage{
xlerate(path, inputs, outputs, check = TRUE)
}
\arguments{
\item{path}{Path to the xlsx file}

\item{inputs}{A \code{\link{xlerate_ref}} object with input references}

\item{outputs}{A \code{\link{xlerate_ref}} object with output references}

\item{check}{Logical, indicating if the workbook should be
recomputed to check that we agree with Excel}
}
\description{
Create a function from an excel workbook
}
