% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlerate.R
\name{xlerate_ref}
\alias{xlerate_ref}
\title{Create a reference to inputs or outputs}
\usage{
xlerate_ref(ref, sheet, label = NULL)
}
\arguments{
\item{ref}{A character vector of cell references (can be ranges)}

\item{sheet}{The sheet number to be used wherever a non-sheet
scoped reference is used.}

\item{label}{Optional list of row/column offsets to locate labels}
}
\description{
Create a reference to inputs or outputs
}
